requirements-dev.txt
wheel
nose
exdoc
j2cli
flask
smsframework
pyjwt
requests-mock
